Update calculation of 3 numbers.py

+++ calculation of 3 numbers.py
@@ -1,4 +1,4 @@
 A=int(input())
 B=int(input())
 C=int(input())
-print((A*B)/C)
+print(int((A*B)/C))
